Clean up code and new dataset included
The data now properly points to the sample data for OIT. The map does
not draw the new locations properly.
Bugs: Clicking on a point on the map does not load a page for that
location, choosing to display a location on the map does not properly
highlight it with a blue marker. For some reason, the geolocate icon on
the map page isn't loading. Probably a CSS issue.